Created foldconstants pass
